fix: address additional SonarQube reliability issues in libinterp
- comp-amd64.c: initialize n=0 to prevent garbage value in bad: cleanup path
- das-s800.c: change inst() from static int to static void (never returns a value)
- das-thumb.c: add unreachable return -1 after complete switch coverage
- string.c: cast (v+1) to int before negation to avoid unary minus on unsigned

+++ libinterp/string.c
@@ -74,7 +74,7 @@ r:
 			ss->Srune[v] = r;
 		else
 		if(v == l && v < ss->max) {
-			ss->len = -(v+1);
+			ss->len = -(int)(v+1);
 			ss->Srune[v] = r;
 		}
 		else {
@@ -83,7 +83,7 @@ r:
 			ns = newstring((v + 1 + v/4)*sizeof(Rune));
 			memmove(ns->Srune, ss->Srune, -ss->len*sizeof(Rune));
 			ns->Srune[v] = r;
-			ns->len = -(v+1);
+			ns->len = -(int)(v+1);
 			ns->max /= sizeof(Rune);
 			ss = ns;
 		}
